requirements.txt
tensorflow==2.10.0
streamlit==1.45.1
pillow==9.5.0
numpy==1.26.4
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.6.1
opencv-python==4.11.0.86
